+++ packages/horizon/contracts/mocks/imports.sol
@@ -10,12 +10,14 @@ import "@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol";
 import "@graphprotocol/contracts/contracts/governance/Controller.sol";
 import "@graphprotocol/contracts/contracts/upgrades/GraphProxyAdmin.sol";
 import "@graphprotocol/contracts/contracts/staking/IStaking.sol";
+import "@graphprotocol/contracts/contracts/discovery/ISubgraphNFT.sol";
 
 // Also for toolshed, solidity version in @graphprotocol/contracts does not support overriding public getters
 // in interface file, so we need to amend them here.
 import { IRewardsManager } from "@graphprotocol/contracts/contracts/rewards/IRewardsManager.sol";
 import { IL2Curation } from "@graphprotocol/contracts/contracts/l2/curation/IL2Curation.sol";
 import { IEpochManager } from "@graphprotocol/contracts/contracts/epochs/IEpochManager.sol";
+import { IGNS } from "@graphprotocol/contracts/contracts/discovery/IGNS.sol";
 
 interface IRewardsManagerToolshed is IRewardsManager {
     function subgraphService() external view returns (address);
@@ -27,4 +29,8 @@ interface IL2CurationToolshed is IL2Curation {
 
 interface IEpochManagerToolshed is IEpochManager {
     function epochLength() external view returns (uint256);
+}
+
+interface IGNSToolshed is IGNS {
+    function subgraphNFT() external view returns (address);
 }