+++ packages/horizon/ignition/modules/periphery/periphery.ts
@@ -5,6 +5,7 @@ import { ethers } from 'ethers'
 import ControllerModule, { MigrateControllerDeployerModule } from './Controller'
 import CurationModule, { MigrateCurationDeployerModule } from './Curation'
 import EpochManagerModule, { MigrateEpochManagerModule } from './EpochManager'
+import GNSModule, { MigrateL2GNSModule } from './GNS'
 import GraphProxyAdminModule, { MigrateGraphProxyAdminModule } from './GraphProxyAdmin'
 import GraphTokenGatewayModule, { MigrateGraphTokenGatewayModule } from './GraphTokenGateway'
 import GraphTokenModule, { MigrateGraphTokenModule } from './GraphToken'
@@ -16,6 +17,7 @@ export default buildModule('GraphHorizon_Periphery', (m) => {
 
   const { EpochManager, EpochManagerImplementation } = m.useModule(EpochManagerModule)
   const { L2Curation, L2CurationImplementation } = m.useModule(CurationModule)
+  const { L2GNS, L2GNSImplementation, SubgraphNFT } = m.useModule(GNSModule)
   const { RewardsManager, RewardsManagerImplementation } = m.useModule(RewardsManagerModule)
   const { L2GraphTokenGateway, L2GraphTokenGatewayImplementation } = m.useModule(GraphTokenGatewayModule)
   const { L2GraphToken, L2GraphTokenImplementation } = m.useModule(GraphTokenModule)
@@ -26,6 +28,7 @@ export default buildModule('GraphHorizon_Periphery', (m) => {
   m.call(Controller, 'setContractProxy', [ethers.keccak256(ethers.toUtf8Bytes('GraphTokenGateway')), L2GraphTokenGateway], { id: 'setContractProxy_GraphTokenGateway' })
   m.call(Controller, 'setContractProxy', [ethers.keccak256(ethers.toUtf8Bytes('GraphProxyAdmin')), GraphProxyAdmin], { id: 'setContractProxy_GraphProxyAdmin' })
   m.call(Controller, 'setContractProxy', [ethers.keccak256(ethers.toUtf8Bytes('Curation')), L2Curation], { id: 'setContractProxy_L2Curation' })
+  m.call(Controller, 'setContractProxy', [ethers.keccak256(ethers.toUtf8Bytes('GNS')), L2GNS], { id: 'setContractProxy_L2GNS' })
 
   return {
     Controller,
@@ -33,6 +36,9 @@ export default buildModule('GraphHorizon_Periphery', (m) => {
     EpochManagerImplementation,
     L2Curation,
     L2CurationImplementation,
+    L2GNS,
+    L2GNSImplementation,
+    SubgraphNFT,
     GraphProxyAdmin,
     L2GraphToken,
     L2GraphTokenImplementation,
@@ -45,6 +51,7 @@ export default buildModule('GraphHorizon_Periphery', (m) => {
 
 export const MigratePeripheryModule = buildModule('GraphHorizon_Periphery', (m) => {
   const { L2CurationProxy: L2Curation, L2CurationImplementation } = m.useModule(MigrateCurationDeployerModule)
+  const { L2GNS, L2GNSImplementation, SubgraphNFT } = m.useModule(MigrateL2GNSModule)
   const { RewardsManagerProxy: RewardsManager, RewardsManagerImplementation } = m.useModule(MigrateRewardsManagerDeployerModule)
   const { Controller } = m.useModule(MigrateControllerDeployerModule)
   const { GraphProxyAdmin } = m.useModule(MigrateGraphProxyAdminModule)
@@ -53,12 +60,14 @@ export const MigratePeripheryModule = buildModule('GraphHorizon_Periphery', (m) 
   const { L2GraphTokenGateway } = m.useModule(MigrateGraphTokenGatewayModule)
 
   // Load these contracts so they are available in the address book
-
   return {
     Controller,
     EpochManager,
     L2Curation,
     L2CurationImplementation,
+    L2GNS,
+    L2GNSImplementation,
+    SubgraphNFT,
     GraphProxyAdmin,
     L2GraphToken,
     L2GraphTokenGateway,