+++ packages/subgraph-service/tasks/deploy.ts
@@ -74,6 +74,9 @@ task('deploy:protocol', 'Deploy a new version of the Graph Protocol Horizon cont
           subgraphServiceProxyAddress: proxiesDeployment.Transparent_Proxy_SubgraphService.target as string,
           subgraphServiceProxyAdminAddress: proxiesDeployment.Transparent_ProxyAdmin_SubgraphService.target as string,
           graphTallyCollectorAddress: horizonDeployment.GraphTallyCollector.target as string,
+          gnsProxyAddress: horizonDeployment.Graph_Proxy_L2GNS.target as string,
+          gnsImplementationAddress: horizonDeployment.Implementation_L2GNS.target as string,
+          subgraphNFTAddress: horizonDeployment.SubgraphNFT.target as string,
         },
       }),
     })